requirements.txt
# MTR Normalization System Requirements

# Core dependencies
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2  # Excel file handling
xlsxwriter==3.1.9

# AI and ML
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
sentence-transformers==2.2.2
transformers==4.36.2

# Vector database
pinecone-client==3.0.0
chromadb==0.4.22  # Alternative to Pinecone for local testing

# Web scraping and search
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
selenium==4.16.0  # For dynamic content
googlesearch-python==1.2.3

# Async and concurrency
asyncio==3.4.3
aiofiles==23.2.1
tenacity==8.2.3  # Retry logic

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
click==8.1.7  # CLI
rich==13.7.0  # Beautiful terminal output
tqdm==4.66.1  # Progress bars

# Logging and monitoring
loguru==0.7.2
sentry-sdk==1.39.1  # Error tracking (optional)

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-mock==3.12.0

# Russian language support
pymorphy2==0.9.1
nltk==3.8.1

# Document parsing
unstructured==0.11.6
pypdf==3.17.4
python-docx==1.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
